Cast average days to int.

+++ src/views/EmployeePage.java
@@ -378,7 +378,7 @@ public class EmployeePage extends JFrame {
 				if(SqlDateFormatHelper.isValidSqlDateString(startDateText) && SqlDateFormatHelper.isValidSqlDateString(endDateText)){
 					Calendar startDate = SqlDateFormatHelper.SQLDateStringToCalendar(startDateText);
 					Calendar endDate = SqlDateFormatHelper.SQLDateStringToCalendar(endDateText);
-					double averageStay = _queryManager.getAverageStayDuration(startDate, endDate);
+					int averageStay = (int)_queryManager.getAverageStayDuration(startDate, endDate);
 					stay_duration_label.setText(String.valueOf(averageStay) + " days");
 				}
 			}
